telegram_bot/requirements.txt
# --- Core Framework ---
# Use version 21+ for Python 3.11/3.12 compatibility
python-telegram-bot>=21.0

# --- AI & LLM ---
google-generativeai==0.3.2

# --- Utilities ---
python-dotenv==1.0.0
aiohttp==3.9.1